# Optimizer CPU offload for single GPU training