--- a/torchao/prototype/low_bit_optim/__init__.py
+++ b/torchao/prototype/low_bit_optim/__init__.py
@@ -1,3 +1,3 @@
 from .adam import Adam8bit, Adam4bit, AdamFp8
 from .adamw import AdamW8bit, AdamW4bit, AdamWFp8
-from .cpu_offload import CPUOffloadOptimizer
+from .cpu_offload import CPUOffloadOptimizer, CPUOffloadOptimizerv2