## Background

Currently there is no simple way to do optimizer CPU offload for single GPU training, although such feature exists for FSDP. DeepSpeed [ZeRO-Offload](https://www.deepspeed.ai/tutorials/zero-offload/) can work with single GPU, but it requires installing DeepSpeed, which can be complicated, and major changes to training loop (not convenient to switch between DeepSpeed and non-DeepSpeed).

Optimizer memory footprint is the largest in a training setup (2x model size for plain Adam), thus offloading optimizer to CPU would be greatly beneficial.

Below is a copy of optimizer CPU offload README

## Optimizer CPU Offload

This folder also implements optimizer CPU offload (i.e. ZeRO-Offload) for single GPU training. For multi-GPU training, you can use FSDP's built-in CPU offload.

```python
import torch
from torchao.prototype.low_bit_optim import CPUOffloadOptimizer

model = ...

# only offload optimizer state
optim = CPUOffloadOptimizer(model.parameters(), torch.optim.AdamW, fused=True)

# offload optimizer state AND gradients
optim = CPUOffloadOptimizer(model.parameters(), torch.optim.AdamW, offload_gradients=True, fused=True)
```

This will reduce GPU memory usage by optimizer state size, and additionally gradient size if `offload_gradients=True`. `CPUOffloadOptimizer` can wrap any base optimizer.

For saving and loading `CPUOffloadOptimizer`, it is important that you load model's weights BEFORE creating the optimizer, since we create a CPU copy of the parameters inside `CPUOffloadOptimizer.__init__()`. (TODO: we might want to have a method to synchronize CUDA and CPU params in either direction CPU->CUDA and CUDA->CPU, in case they are out of sync.)

```python
ckpt = torch.load("checkpoint.pth")

model = ...
model.load_state_dict(ckpt["model"])

optim = CPUOffloadOptimizer(model.parameters(), torch.optim.AdamW, fused=True)
optim.load_state_dict(ckpt["optim"])
```

NOTE:
- Since the optimizer step is done on CPU, it is highly recommended to use a fast CPU optimizer, such as `torch.optim.AdamW(fused=True)` (requires PyTorch 2.4). For other optimizers, you can try `torch.compile()` their optimizer step.
- To minimize the amount of CPU<->GPU data transfer, we keep a copy of parameters and pre-allocate gradients memory on CPU. Therefore, expect your RAM usage to increase by 2x model size + optimizer state (which is 2x model size for Adam).
- It is recommended NOT to `torch.compile()` your whole model when `CPUOffloadOptimizer` is used, as it prevents us from interleaving gradient device-to-host transfer with backward pass. To minimize such impact, you can compile parts of your model separately. See [#584](https://github.com/pytorch/ao/pull/584) for more information.
- CPU optimizer step is often the bottleneck when optimizer CPU offload is used. To minimize the slowdown, it is recommended to (1) do full BF16 training (instead of AMP), so that parameters, gradients, and optimizer states are in BF16; and (2) give GPU more work per optimizer step (e.g. larger batch size with activation checkpointing, gradient accumulation).
- `offload_gradients=True` is not compatible with gradient accumulation, since we clear gradients on GPU every backward pass.
- Gradient clipping is currently not supported.

Benchmark done for `timm/vit_giant_patch14_dinov2.lvd142m` (1.1B params), eager mode, full BF16 training, activations checkpointing, batch size 32, on 4070Ti SUPER (16GB VRAM), Ryzen 5600, DDR4 RAM. DeepSpeed is untuned.

Adam offload           | Time per step | Max memory
-----------------------|---------------|------------
None                   | 1.27s/it      | 9.82 GB
DeepSpeed ZeRO-Offload | 3.13s/it      | 6.85 GB
ao                     | 1.52s/it      | 5.24 GB
ao (offload gradients) | 1.53s/it      | 4.01 GB

Ablations on AMP and `torch.compile()`

Training config     | Adam offload | Time per step | Max memory
--------------------|--------------|---------------|------------
Full BF16, compiled | None         | 1.18s/it      | 9.90 GB
Full BF16, compiled | ao           | 1.75s/it      | 5.33 GB
BF16 AMP, eager     | None         | OOM           | OOM
BF16 AMP, eager     | ao           | 2.18s/it      | 9.90 GB

## Implementation details

Keep a copy of params on CPU. After backward pass, copy gradients from GPU to CPU (optionally deallocate GPU gradients). Do optimizer step on CPU. Copy updated gradients from CPU to GPU.

To hide CPU <-> GPU data movements, interleave grad device->host with backward, and interleave param host->device with CPU optim step. Also start CPU optim step as soon as CPU is free (i.e. after launching all backward kernels) -> some interleaving of backward and CPU optim step. The following trace illustrates the strategy.

![image](https://github.com/user-attachments/assets/91708eed-b5b6-40a3-ab98-144faf8fe9a3)

Two interesting observations:

**1. `torch.compile()` prevents overlapping grad D2H with backward**. Probably because compiled backward will launch/queue all backward kernels at once, so waiting for current stream will last until all backwards finish. Trace with torch.compile()

![image](https://github.com/user-attachments/assets/b65d74bb-4bad-4cd9-8e19-b0e97159732b)

One way to mitigate this is to compile parts of the model separately, so on the host side, backwards are launched as K groups of kernels, then we can start grad D2H in-between. (haven't tried it, just an idea. maybe still not possible). This would also reduce kernel launch overhead, which helps CPU Adam to start even earlier.

**2. Fused CPU Adam is much faster in BF16 than in FP32**. Trace with BF16 AMP (params, grads, optimizer states are in FP32), optim step time increases from ~700ms -> 1200ms.

![image](https://github.com/user-attachments/assets/98de895d-9368-4427-9f3d-6372b5c95f73)

Time for CUDA optim: CUDA forward time + CUDA backward time + CUDA optim time.
Time with CPU offload optim: CUDA forward time + CPU time to launch backward + CPU optim time.

If CPU optim time is super fast (assume zero), the upper bound will be CUDA forward time + CUDA backward time + param H2D time (though we can further hide H2D time with forward). So optim CPU offload MAY even be faster than optim on GPU.

My setup is Ryzen 5600 with dual-channel DDR4. Using CPUs with AVX-512 support (Ryzen 4 and later or server CPUs) and DDR5 (or 6-channel DDR4 on servers) would probably be faster.