@gau-nernst 2024-08-01
@awgu Thank you for your feedback.

For "overlap gradient D2H copies with backward", it probably cannot be done without intrusive change to the training code? Perhaps something like [this](https://pytorch.org/tutorials/intermediate/optimizer_step_in_backward_tutorial.html) can help with the overlapping (i.e. once we finish accumulating gradient for a param, we start moving it to CPU, while still doing backward for other params. Since optim step on CPU is blocking, we can only do optim step once all gradients are copied to CPU?). 

For "H2D copies with forward", how can I do this? I have been reading [this](https://pytorch.org/tutorials/intermediate/pinmem_nonblock.html), and it says that I need to use a separate CUDA stream for data transfer to overlap with computation. So it means that I have to somehow synchronize the CPU->CUDA transfer (which will be in a separate CUDA stream) before that param is needed for forward? Perhaps some kind of forward hook? (again, intrusive changes to the training code T.T)