@awgu 2024-08-01
I agree once you want to overlap, it becomes quite intrusive 😢 .

The post-accumulate-grad hook could be a good point to run the D2H copy for gradients, but again like you said, without a separate CUDA stream, that copy is not going to overlap. You will mainly be moving the same kernel that would happen in `optimizer.step()` into the backward.

I do not have a good solution for how to do the overlap in a non-intrusive way. From what I have seen, it is hard to do this kind of overlap without some kind of `nn.Module` level API since that gives you good points to hook into the forward and backward.