@gau-nernst 2024-08-01
Add a poor man's attempt at interleaving grad D2H with backward. seems to work! (you can check the latest changes). speed improves from 1.2s/it to 1.0s/it.

Before (blue is backward kernels, red is copy kernels)
<img width="810" alt="image" src="https://github.com/user-attachments/assets/ee582213-5505-40b9-a2c9-82e7b875ade9">

After
<img width="795" alt="image" src="https://github.com/user-attachments/assets/040ed5a2-b998-4bbd-8a59-54d222e6c350">

in the 2nd image, one thing concerns me is the backward kernels (blue) ends after copy kernels (red). Maybe some bugs with torch profiler? (I'm using `torch.profiler.profile`, the trace is obtained from the `benchmarks/benchmark_low_bit_adam.py` script). Loss curve seems ok.

The profiling trace also reveals that CPU optimizer step is the bottleneck, which we won't be able to hide.

<img width="1134" alt="image" src="https://github.com/user-attachments/assets/4e6983be-b1df-4472-a0d0-4e3bd5bbcc06">

So I think in this workload, I should try to feed more work to GPU (e.g. larger batch size, gradient accumulation...).