@awgu 2024-08-01
Oh, I did not realize that the profiler was profiling so many steps. I think it okay to just profile 1 step? I am okay with anyway for you to share it.