@gau-nernst 2024-08-01
Here it is

[optim_cpu_offload_d2h_overlap.tar.gz](https://github.com/user-attachments/files/16458120/optim_cpu_offload_d2h_overlap.tar.gz)

For interleaving H2D param copy with forward, I'm thinking of using `nn.Module.register_forward_pre_hook()`. But even with the forward hook, it is tricky to know which params under that module should be synchronized. Maybe we only synchronize the immediate `nn.Parameter` (there is no direct API for this I think, but using `.named_parameters()` and check for no prefix should be ok).