@gau-nernst 2024-08-05
I sent the DeepSpeed trace on Discord. Two main reasons (1) DeepSpeed CPU Adam is slower than PyTorch fused Adam, (2) They don't interleave data transfer as well as CPU optim step (might be because I didn't set the config correctly).