@gau-nernst 2024-08-05
Regarding gradient clipping, I thought about it too. The biggest improvement is actually from overlap CPU Adam with backward (i.e. start CPU Adam as soon as host finish launching all backward kernels). We can still move grad D2H during backward (help with hiding data transfer + offload gradients), but CPU Adam can only start when all gradients are present on CPU to do gradient clipping. Even if we do gradient clipping on GPU, CPU Adam still needs to wait for all gradients to be available (i.e. backward finish).

Probably good to add a note that this CPU offload optimizer doesn't support gradient clipping at the moment. We can add support for it in a future PR.