@Theodotus1243 2024-09-21
How to use `CPUOffloadOptimizer` with `LRScheduler`
As it has check
```
# Attach optimizer
        if not isinstance(optimizer, Optimizer):
            raise TypeError(f'{type(optimizer).__name__} is not an Optimizer')
```

And 
```
class CPUOffloadOptimizer:
    def __init__(
```