@gau-nernst 2024-09-21
Hi @Theodotus1243, you have to manually set the LR, since built-in PyTorch's LRScheduler will enforce the optimizer to be an `torch.optim.Optimizer` subclass as you have discovered. Something like this

https://github.com/pytorch/ao/blob/0bdde92114b470823aa24725bf3b0811e980c8ce/benchmarks/benchmark_low_bit_adam.py#L256-L261

The reason I don't want to make `CPUOffloadOptimizer` to be a `torch.optim.Optimizer` subclass is that it doesn't seem right: `CPUOffloadOptimizer` itself doesn't hold the params and buffers, it delegates to the base optimizer class, and only hold a list of base optimizers.

Hope it clarifies the problem. I think we can add this caveat to doc.