@bghira 2024-09-26
i think it should not be referred to as a drop-in replacement then @gau-nernst and as it is, having a method that sets the lr isn't too much to ask for, i hope?